Fix compilation warning on using u64
initvals.c:401: warning: format ‘%016llx’ expects type ‘long long unsigned int’, but argument 2 has type ‘u64’

+++ initvals.c
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 typedef uint32_t u32;
-typedef uint64_t u64;
+typedef long long unsigned int u64;
 
 /*
  * compile with -DATHEROS if you want to use the upstream Atheros initvals
